Add UI functionality
Makes the Rock Paper Scissors game playable by clicking
browser buttons.

Control flow is now as follows:
- User selects rock, paper, or scissors.
- playRound() is called. After the score and round
number are modified and displayed in-browser,
playRound() calls checkWinCondition().
- if checkWinCondition() determines the game is over,
endGame() is called. This disables the r, p, s
buttons and adds a "reset game" button to the DOM.
- resetGame() clears all game history displayed in-browser,
and resets the score and round counters.